requirements.txt
torch
torchvision
opencv-python
numpy
pandas
scikit-learn
matplotlib
streamlit
tqdm
Pillow
